adjust cnp
