+++ autoemulate/emulators/__init__.py
@@ -1,3 +1,4 @@
+from .cnp import CNP
 from .gaussian_process import GaussianProcess
 from .gaussian_process_mogp import GaussianProcessMOGP
 from .gradient_boosting import GradientBoosting
@@ -24,4 +25,5 @@ MODEL_REGISTRY = {
         module="RadialBasisFunctionsNetwork"
     ),
     NeuralNetSk().model_name: NeuralNetSk(),
+    CNP().model_name: CNP(),
 }
